DAY 8/content_creator/requirements.txt
langgraph
langchain-google-genai
python-dotenv
